requirements.txt
click==7.1.27.1.2
numpy==1.19.1
pytest==6.0.1
PyYaml==5.3.1
rich==5.1.0
tqdm==4.48.2
